V1.2.2.1

+++ org.eclipse.paho.client.mqttv3/src/main/java/org/eclipse/paho/client/mqttv3/internal/wire/MultiByteInteger.java
@@ -2,13 +2,13 @@
  * Copyright (c) 2009, 2014 IBM Corp.
  *
  * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
+ * are made available under the terms of the Eclipse Public License v2.0
  * and Eclipse Distribution License v1.0 which accompany this distribution. 
  *
  * The Eclipse Public License is available at 
- *    http://www.eclipse.org/legal/epl-v10.html
+ *    https://www.eclipse.org/legal/epl-2.0
  * and the Eclipse Distribution License is available at 
- *   http://www.eclipse.org/org/documents/edl-v10.php.
+ *   https://www.eclipse.org/org/documents/edl-v10.php
  *
  * Contributors:
  *    Dave Locke - initial API and implementation and/or initial documentation
@@ -20,14 +20,14 @@ package org.eclipse.paho.client.mqttv3.internal.wire;
  * specification.
  */
 public class MultiByteInteger {
-	private long value;
+	private int value;
 	private int length;
 	
-	public MultiByteInteger(long value) {
+	public MultiByteInteger(int value) {
 		this(value, -1);
 	}
 	
-	public MultiByteInteger(long value, int length) {
+	public MultiByteInteger(int value, int length) {
 		this.value = value;
 		this.length = length;
 	}
@@ -42,7 +42,7 @@ public class MultiByteInteger {
 	/**
 	 * @return the value of this MBI.
 	 */
-	public long getValue() {
+	public int getValue() {
 		return value;
 	}
 }
